refactor(player): refactor player module
Linked to #33
